Actualizar index.html
adjust the skill tags animation to appear a bit later than what you currently have it in relation to the scroll bar position.​​​​​​​​​​​​​​​​ also add a key to identify what each color means. Also expand the project cards when clicked on with an animation to bring them to the top front of the screen and expand the details on that project. Eventually I hope to include images to do an image carrousel when expanded too and add independent skill tags to each project in the expanded view